+++ domain/pricing.go
@@ -104,3 +104,20 @@ func FormatPricingCacheKey(a, b string) string {
 	sb.WriteString(b)
 	return sb.String()
 }
+
+type PricingWorker interface {
+	// UpdatePrices updates prices for the given base denoms asyncronously.
+	// Returns a channel that will be closed when the update is completed.
+	// Propagates the results to the listeners.
+	UpdatePricesAsync(height uint64, baseDenoms map[string]struct{})
+
+	// RegisterListener registers a listener for pricing updates.
+	RegisterListener(listener PricingUpdateListener)
+
+	// IsProcessing returns true if the worker is processing a pricing update.
+	IsProcessing() bool
+}
+
+type PricingUpdateListener interface {
+	OnPricingUpdate(ctx context.Context, height int64, pricesBaseQuoteDenomMap map[string]map[string]any, quoteDenom string) error
+}